pyproject.toml
[project]
name = "flypi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.11",
    "fastapi[standard]>=0.115.8",
    "python-dotenv>=1.0.1",
]

[project.scripts]
flypi = "flypi.__main__:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
]
